move inline stuff to its submodule
