+++ vsexprtools/__init__.py
@@ -1,6 +1,4 @@
 from .exprop import *
 from .funcs import *
-from .manager import *
-from .operators import *
+from .inline import *
 from .util import *
-from .variables import *
